Add files via upload
This file basically takes the dimensions of the canvas and on top of that it tries to build rectangles or squares, with the color schema. As we know all colors are the combination of red , blue , green. IT has different classes being made and those classes use the command line interface to take inputs and display the results.